updates on prices & quantity

+++ Company.java
@@ -13,11 +13,11 @@ public abstract class Company {
 	protected double myMarketShare;
 	protected String myCountry;
 	double myPrice;
-	int myQuantity;
+	double myQuantity;
 	protected boolean typeOfCompany; //True if collusive, false if defector
 
 	// lifeform constructors
-	public Company(double myBudget, ArrayList<Location> myLocation, Color myColor, double myMarketShare, String myCountry, boolean typeOfCompany, double myPrice, int myQuantity) {
+	public Company(double myBudget, ArrayList<Location> myLocation, Color myColor, double myMarketShare, String myCountry, boolean typeOfCompany, double myPrice, double myQuantity) {
 		super();
 		this.myBudget = myBudget;
 		this.myLocations = myLocation;
@@ -25,8 +25,8 @@ public abstract class Company {
 		this.myMarketShare = myMarketShare;
 		this.myCountry = myCountry;
 		this.typeOfCompany = typeOfCompany;
-		this.myPrice = myPrice;
-		this.myQuantity = myQuantity;
+		this.myPrice = myPrice; // Prices are based on 2019 Oil prices adjusted from the local currency to US dollars
+		this.myQuantity = myQuantity; // Quantity starts off the same (ten million barrels) 10000000
 		alive = true;
 	}
